introduces 'stackcore', a separate and complementary code bucket to 'buildercore'. new resource stack logic can grow in here without affecting 'buildercore', especially since the logic is so tightly coupled the project template approach of defining resources.
